requirements.txt
streamlit
tensorflow
numpy
pillow
pandas
google-generativeai
